graph: hand the rest of the tree algorithms to navis-fastcore
The fallbacks went away last commit but the fast paths were only ever written
for the primitives that existed at the time. These are the ones that were still
igraph or networkx underneath because fastcore had nothing to call, and now
does. Floor moves to 0.10.0, which is the first release carrying them.

`betweeness_centrality` was Brandes over a graph built per call. On a tree the
answer is closed form -- a node lies on the path between exactly those pairs
separated by it, which is descendants times ancestors -- so `betweenness` gets
it in one pass. With `from_` it is `descendant_counts` against those sources,
which also drops the post-hoc filter that used to sit on top of the full
matrix: counting only the sources asked for is the same number, arrived at
without computing the ones thrown away.

`node_label_sorting` drove its sort with a directed breaks-by-breaks geodesic
matrix -- 4.5 GB on a 71k-node skeleton -- where the key is subtree height plus
a difference of root distances, both O(N). The two are the same quantity;
`tests/test_node_label_sorting.py` pins the new one against the old
formulation. Children are enumerated in node-table order now rather than in the
insertion order of a networkx adjacency set, so ties break deterministically.
Where they tie the order shifts.

`collapse_nodes` was building an igraph contraction mapping of vertex *indices*
and writing node IDs into it. That only held together while IDs happened to run
1..N; given the 7e17-range IDs segmentation backends hand out, igraph tried to
reserve a vector sized by the ID and died. `contract_nodes` works in ID space,
which is where the mapping was all along.

`_igraph_to_sparse` is gone. `propagate_labels` branched on neuron type to pick
between it and a fastcore adjacency, but the mesh half did not need igraph
either -- a Mesh's `.igraph` is built from `mesh_unique_edges`, so it was
constructing a graph to read its own edges back out. One `_sparse_adjacency`
handles both.

The rest are direct swaps: `find_main_branchpoint` and `split_into_fragments`
onto `longest_paths`, `_cut_igraph` onto `descendants`, `reroot_skeleton` onto
`reroot`, `rewire_skeleton` onto `minimum_spanning_tree` + `parents_from_edges`,
`edges2neuron` onto `parents_from_edges`, `cell_body_fiber` and `_plot_subway`
onto `paths_to_root`, `Skeleton.is_tree` onto `has_cycles`. Nine copies of the
parent-distance block collapse onto `morpho.mmetrics.parent_dist`.

Two fixes come in from fastcore rather than from here, and are in the changelog
as such: `geodesic_matrix(directed=True)` no longer reports a coincident child
as reachable from its parent, and weighted segment lengths are no longer one
edge too long -- they now sum to exactly the neuron's cable length.

`geodesic_matrix` pins float32 on the mesh-with-extra-edges branch. 0.10.0
takes the distances' width from the weights', and `mesh_unique_edges` hands
back float64 because trimesh vertices are, which would have made that one
branch of the function return float64 while the other two could not follow.

Mesh edges are cast to int64 before they seed trimesh's cache: fastcore returns
uint32 under its own dtype rules, trimesh's `edges_unique` is int64, and this is
a trimesh cache key that third-party code reads and may do signed arithmetic on.

+++ navis/utils/misc.py
@@ -544,7 +544,11 @@ def mesh_unique_edges(x, return_lengths=False, extra_edges=True):
         else:
             edges, index, inverse = res
 
-        # Update cache (mirrors what trimesh's `edges_unique` populates)
+        # Update cache (mirrors what trimesh's `edges_unique` populates). N.B. the
+        # cast: fastcore hands back `uint32` (a node ID under its dtype rules) but
+        # trimesh's own `edges_unique` is `int64`, and this is a *trimesh* cache key
+        # that third-party code reads and may do signed arithmetic on.
+        edges = edges.astype(np.int64)
         mesh._cache["edges_unique"] = edges
         mesh._cache["edges_unique_idx"] = index
         mesh._cache["edges_unique_inverse"] = inverse
